Implement demo rating service

+++ rating/src/main/java/diploma/autos/rating/repositories/EngineRepository.java
@@ -5,9 +5,7 @@ import diploma.autos.rating.entities.Engine;
 import diploma.autos.rating.entities.EnginePK;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
-
 
 public interface EngineRepository extends CrudRepository<Engine, EnginePK> {
-    Engine findByBrandAndModelAndVolume(String brand, String model, int volume);
+    Engine findByBrandAndModelAndVolume(String brand, String model, Double volume);
 }
